requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.4
chess==1.6.1
click==8.0.1
Cython==0.29.24
Flask==2.0.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysql-connector-python==8.0.26
numpy==1.21.2
pandas==1.3.2
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
six==1.16.0
urllib3==1.26.6
Werkzeug==2.0.1
pyyaml==5.4.1
